+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.0.4 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- maximum_backoff -> max_backoff
+- on_exception no longer may be a list - only tuple is allowed as an iterable
+- add documentation/readme
+- add examples.py
 
 
 ## 0.0.3 (2025-11-17)